refactor: make all needed variables from start
The code was messy because it had some wierd
functions which were confusing and specific.
I removed that by adding a simple function
which returns a map with all needed variables,
instead of having to use about 3 functions over
and over again to parse the request and get
the necessary variables.